consolidate vote extensions to the oracle module
